+++ yt_shared/src/yt_shared/rabbit/rabbit_config.py
@@ -5,10 +5,12 @@ from aio_pika import ExchangeType
 INPUT_QUEUE = 'input.q'
 ERROR_QUEUE = 'error.q'
 SUCCESS_QUEUE = 'success.q'
+INFO_QUEUE = 'info.q'
 
 INPUT_EXCHANGE = 'input.dx'
 SUCCESS_EXCHANGE = 'success.dx'
 ERROR_EXCHANGE = 'error.dx'
+INFO_EXCHANGE = 'info.dx'
 
 
 def get_rabbit_config() -> dict[str, list[dict[str, Any]]]:
@@ -17,6 +19,7 @@ def get_rabbit_config() -> dict[str, list[dict[str, Any]]]:
             {'name': INPUT_QUEUE, 'auto_delete': False, 'durable': True},
             {'name': ERROR_QUEUE, 'auto_delete': False, 'durable': True},
             {'name': SUCCESS_QUEUE, 'auto_delete': False, 'durable': True},
+            {'name': INFO_QUEUE, 'auto_delete': False, 'durable': True},
         ],
         'exchanges': [
             {
@@ -37,10 +40,17 @@ def get_rabbit_config() -> dict[str, list[dict[str, Any]]]:
                 'durable': True,
                 'type': ExchangeType.DIRECT.value,
             },
+            {
+                'name': INFO_EXCHANGE,
+                'auto_delete': False,
+                'durable': True,
+                'type': ExchangeType.DIRECT.value,
+            },
         ],
         'queue_bindings': {
             INPUT_QUEUE: [{'exchange_name': INPUT_EXCHANGE}],
             ERROR_QUEUE: [{'exchange_name': ERROR_EXCHANGE}],
             SUCCESS_QUEUE: [{'exchange_name': SUCCESS_EXCHANGE}],
+            INFO_QUEUE: [{'exchange_name': INFO_EXCHANGE}],
         },
     }