+++ yt_shared/src/yt_shared/schemas/url.py
@@ -16,3 +16,4 @@ class URL(StrictRealBaseModel):
     ack_message_id: int
     save_to_storage: bool
     download_media_type: DownMediaType
+    playlist: bool = False